chore(profiling): Lift PerformanceDuration component
Previously, the PerformanceDuration component was only being used in the
performance product. With the introduction of profiling, PerformanceDuration
will be used in more places as the default component for displaying duration
values. This change lifts the component into the top level components to avoid
profiling having to import from performance views.
